// examples/prisma-8-demo/src/prisma/contract.d.ts
export type CodecTypes = PgTypes & PgVectorTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes> &
  PgVectorQueryOperationTypes<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Bug: {
      readonly severity: CodecTypes['pg/text@1']['output'];
      readonly stepsToRepro: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Feature: {
      readonly priority: CodecTypes['pg/text@1']['output'];
      readonly targetRelease: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Post: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly embedding: Vector<1536> | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly impressionCount: CodecTypes['pg/int8@1']['output'] | null;
      readonly priority: 0 | 1 | 2;
      readonly reachScore: CodecTypes['pg/unboundedint@1']['output'] | null;
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly viewCount: CodecTypes['pg/int8number@1']['output'] | null;
    };
    readonly PostTag: {
      readonly postId: CodecTypes['pg/uuid@1']['output'];
      readonly tagId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly label: CodecTypes['pg/text@1']['output'];
    };
    readonly Task: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly User: {
      readonly address: AddressOutput | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly displayName: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly kind: 'admin' | 'user';
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Bug: {
      readonly severity: CodecTypes['pg/text@1']['input'];
      readonly stepsToRepro: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Feature: {
      readonly priority: CodecTypes['pg/text@1']['input'];
      readonly targetRelease: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Post: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly embedding: CodecTypes['pg/vector@1']['input'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly impressionCount: CodecTypes['pg/int8@1']['input'] | null;
      readonly priority: 0 | 1 | 2;
      readonly reachScore: CodecTypes['pg/unboundedint@1']['input'] | null;
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly viewCount: CodecTypes['pg/int8number@1']['input'] | null;
    };
    readonly PostTag: {
      readonly postId: CodecTypes['pg/uuid@1']['input'];
      readonly tagId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly label: CodecTypes['pg/text@1']['input'];
    };
    readonly Task: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly User: {
      readonly address: AddressInput | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly displayName: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly kind: 'admin' | 'user';
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly bug: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly severity: CodecTypes['pg/text@1']['output'];
      readonly stepsToRepro: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly feature: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly priority: CodecTypes['pg/text@1']['output'];
      readonly targetRelease: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly post: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly embedding: Vector<1536> | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly impressionCount: CodecTypes['pg/int8@1']['output'] | null;
      readonly priority: 0 | 1 | 2;
      readonly reachScore: CodecTypes['pg/unboundedint@1']['output'] | null;
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
      readonly viewCount: CodecTypes['pg/int8number@1']['output'] | null;
    };
    readonly post_tag: {
      readonly postId: CodecTypes['pg/uuid@1']['output'];
      readonly tagId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly tag: {
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly label: CodecTypes['pg/text@1']['output'];
    };
    readonly task: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly user: {
      readonly address: CodecTypes['pg/jsonb@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly displayName: CodecTypes['pg/text@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly kind: 'admin' | 'user';
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly bug: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly severity: CodecTypes['pg/text@1']['input'];
      readonly stepsToRepro: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly feature: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly priority: CodecTypes['pg/text@1']['input'];
      readonly targetRelease: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly post: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly embedding: CodecTypes['pg/vector@1']['input'] | null;
      readonly expiresAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly impressionCount: CodecTypes['pg/int8@1']['input'] | null;
      readonly priority: 0 | 1 | 2;
      readonly reachScore: CodecTypes['pg/unboundedint@1']['input'] | null;
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
      readonly viewCount: CodecTypes['pg/int8number@1']['input'] | null;
    };
    readonly post_tag: {
      readonly postId: CodecTypes['pg/uuid@1']['input'];
      readonly tagId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly tag: {
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly label: CodecTypes['pg/text@1']['input'];
    };
    readonly task: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly user: {
      readonly address: CodecTypes['pg/jsonb@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly displayName: CodecTypes['pg/text@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly kind: 'admin' | 'user';
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;